task/cleanup-url-utils * move rootPath, buildClassicConfigUrl, buildClassicInputUrl, buildClassicBuildUrl into UrlBuilder namespace
