+++ blueocean-dashboard/src/main/js/components/PipelinePage.jsx
@@ -4,7 +4,7 @@ import Extensions from '@jenkins-cd/js-extensions';
 import { ExpandablePath, Page, TabLink, WeatherIcon } from '@jenkins-cd/design-language';
 import { AppConfig, ContentPageHeader, i18nTranslator, logging, NotFound, Paths, Security } from '@jenkins-cd/blueocean-core-js';
 import { Icon } from '@jenkins-cd/design-language';
-import { buildClassicConfigUrl, UrlBuilder } from '@jenkins-cd/blueocean-core-js';
+import { UrlBuilder } from '@jenkins-cd/blueocean-core-js';
 import { documentTitle } from './DocumentTitle';
 import { observer } from 'mobx-react';
 import { observable, action } from 'mobx';
@@ -17,7 +17,7 @@ const classicConfigLink = pipeline => {
     let link = null;
     if (Security.permit(pipeline).configure()) {
         link = (
-            <a href={buildClassicConfigUrl(pipeline)} title={webTranslate('toast.configure', { defaultValue: 'Configure' })} target="_blank">
+            <a href={UrlBuilder.buildClassicConfigUrl(pipeline)} title={webTranslate('toast.configure', { defaultValue: 'Configure' })} target="_blank">
                 <Icon size={24} icon="ActionSettings" style={{ verticalAlign: 'baseline' }} />
             </a>
         );
